Address PR review comments
