+++ proto/regen/ecocredit/v1alpha1/events.proto
@@ -67,9 +67,9 @@ message EventRetire {
   // units is the decimal number of credits that have been retired.
   string units = 3;
 
-  // location is the location of the beneficiary or buyer of the retiring
+  // location is the location of the beneficiary or buyer of the retired
   // credits. It is a string of the form
   // <country-code>[-<sub-national-code>[-<postal-code>]], with the first two fields
-  // conforming to ISO 3166-2
+  // conforming to ISO 3166-2.
   string location = 4;
 }
