+++ x/ecocredit/go.mod
@@ -3,7 +3,6 @@ module github.com/regen-network/regen-ledger/x/ecocredit
 go 1.15
 
 require (
-	github.com/TheBookPeople/iso3166 v0.0.0-20160919111731-b671ca53a641
 	github.com/btcsuite/btcutil v1.0.3-0.20201208143702-a53e38424cce
 	github.com/cockroachdb/apd/v2 v2.0.2
 	github.com/cosmos/cosmos-sdk v0.43.0-rc0
@@ -11,6 +10,7 @@ require (
 	github.com/gogo/protobuf v1.3.3
 	github.com/gorilla/mux v1.8.0
 	github.com/grpc-ecosystem/grpc-gateway v1.16.0
+	github.com/pariz/gountries v0.0.0-20200430155801-1c6a393df9c7
 	github.com/regen-network/regen-ledger/orm v0.0.0-00010101000000-000000000000
 	github.com/regen-network/regen-ledger/types v0.0.0-00010101000000-000000000000
 	github.com/spf13/cobra v1.1.3
