+++ x/ecocredit/server/msg_server.go
@@ -2,14 +2,12 @@ package server
 
 import (
 	"fmt"
-	"strings"
 
 	"github.com/regen-network/regen-ledger/types"
 
 	"github.com/cockroachdb/apd/v2"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
-	"github.com/TheBookPeople/iso3166"
 
 	"github.com/regen-network/regen-ledger/orm"
 	"github.com/regen-network/regen-ledger/types/math"
@@ -369,48 +367,4 @@ func retire(ctx types.Context, store sdk.KVStore, recipient string, batchDenom b
 		Units:      math.DecimalString(retired),
 		Location:   location,
 	})
-}
-
-// validateLocation checks that the country and region conform to ISO 3166 and
-// the postal code is valid
-func validateLocation(location string) error {
-	country, region, postalCode, err := parseLocation(location)
-	if err != nil {
-		return err
-	}
-	
-	if _, err := iso3166.Decode(country, region, false); err != nil {
-		return err
-	}
-
-	if err := validatePostalCode(postalCode); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func parseLocation(location string) (string, string, string, error) {
-	strings := strings.Split(location, "-")
-	switch numStrings := len(strings); numStrings {
-	case 1:
-		return strings[0], "", "", nil
-	case 2:
-		return strings[0], strings[1], "", nil
-	case 3:
-		return strings[0], strings[1], strings[2], nil
-	default:
-		return "", "", "", fmt.Errorf("Location should have format <country-code>[-<region-code>[-<postal-code>]]")
-	}
-}
-
-// validatePostalCode currently checks that the postal code is not longer than
-// 10 characters, as currently the longest postal codes in the world are 10
-// characters
-func validatePostalCode(postalCode string) error {
-	if len(postalCode) > 10 {
-		return fmt.Errorf("The postal code must not be longer than 10 characters")
-	}
-
-	return nil
 }
